+++ Runtime/Scripts/ErrorCode.cs
@@ -25,6 +25,10 @@ namespace KtxUnity {
         OpenUriFailed,
         LoadingFailed,
         TranscodeFailed,
+        InvalidLayer,
+        InvalidLevel,
+        InvalidFace,
+        InvalidSlice,
     }
 
     public static class ErrorMessage {
@@ -34,9 +38,13 @@ namespace KtxUnity {
             { ErrorCode.UnsupportedVersion, "Only KTX 2.0 is supported" },
             { ErrorCode.UnsupportedFormat, "Unsupported format" },
             { ErrorCode.NotSuperCompressed, "Only super-compressed KTX is supported" },
-            { ErrorCode.OpenUriFailed, "Loading URI failed!" },
-            { ErrorCode.LoadingFailed, "Loading failed!" },
-            { ErrorCode.TranscodeFailed, "Transcoding failed!" },
+            { ErrorCode.OpenUriFailed, "Loading URI failed" },
+            { ErrorCode.LoadingFailed, "Loading failed" },
+            { ErrorCode.TranscodeFailed, "Transcoding failed" },
+            { ErrorCode.InvalidLayer, "Invalid ImageIndex" },
+            { ErrorCode.InvalidLevel, "Invalid MipMapLevel" },
+            { ErrorCode.InvalidFace, "Invalid Face" },
+            { ErrorCode.InvalidSlice, "Invalid Slice" },
         };
 
         public static string GetErrorMessage(ErrorCode code) {